requirements.txt
# 核心功能依赖
srt==3.5.3
openai==1.54.4
httpx==0.27.2
pyyaml==6.0.1

# 异步支持
aiohttp==3.9.3
async-timeout==4.0.3
